Extract reusable invoice uploader panel
